pyproject.toml
[project]
name = "wikimind"
version = "0.1.0"
description = "LLM-powered wiki that maintains itself"
requires-python = ">=3.11"

dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "anthropic>=0.40",
    "python-frontmatter>=1.1",
    "python-slugify>=8.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
pdf = ["pymupdf4llm>=0.0.10"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
wikimind = "wikimind.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
